+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 CHANGELOG
 =========
 
-master
-------
+1.2.4
+-----
 
 Improvements
 
+- Show [solidarity](https://github.com/vshymanskyy/StandWithUkraine) with Ukrainians
 - Show more concise error messages including script exit code #969
 - Configuration (`expression.strip_tailing_zeros`) to strip meaningless zeros after the decimal place #958